backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt==4.0.1
python-multipart>=0.0.6
httpx>=0.24.0
email-validator>=2.0.0
supabase>=2.0.0
Pillow>=10.0.0
